// frontend/src/modules/application/enums/columns-error.ts
const enum ColumnsError {
  All = 'All columns included in the table have been deleted from the spreadsheet',
  OneOrMore = 'One or more columns included in the table have been deleted from the spreadsheet',
  One = 'One column included in the table has been deleted from the spreadsheet',
}

export { ColumnsError };